api-v1/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Client
aiohttp==3.9.1
tenacity==8.2.3

# Google AI
google-generativeai==0.3.2
google-adk==0.1.0

# AWS
boto3==1.34.34

# Authentication
PyJWT==2.8.0

# Environment
python-dotenv==1.0.0
python-multipart==0.0.6
